+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/Metrics.java
@@ -7,10 +7,11 @@ package net.sourceforge.pmd.lang.java.oom;
 
 import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
-import net.sourceforge.pmd.lang.java.oom.api.ClassMetricKey;
+import net.sourceforge.pmd.lang.java.oom.api.ClassMetric;
 import net.sourceforge.pmd.lang.java.oom.api.Metric.Version;
+import net.sourceforge.pmd.lang.java.oom.api.MetricKey;
 import net.sourceforge.pmd.lang.java.oom.api.MetricVersion;
-import net.sourceforge.pmd.lang.java.oom.api.OperationMetricKey;
+import net.sourceforge.pmd.lang.java.oom.api.OperationMetric;
 import net.sourceforge.pmd.lang.java.oom.api.ResultOption;
 
 
@@ -53,7 +54,7 @@ public final class Metrics {
      *
      * @return The value of the metric, or {@code Double.NaN} if the value couln't be computed
      */
-    public static double get(ClassMetricKey key, ASTAnyTypeDeclaration node) {
+    public static double get(MetricKey<ClassMetric> key, ASTAnyTypeDeclaration node) {
         return get(key, node, Version.STANDARD);
     }
 
@@ -68,7 +69,7 @@ public final class Metrics {
      *
      * @return The value of the metric, or {@code Double.NaN} if the value couln't be computed
      */
-    public static double get(ClassMetricKey key, ASTAnyTypeDeclaration node, MetricVersion version) {
+    public static double get(MetricKey<ClassMetric> key, ASTAnyTypeDeclaration node, MetricVersion version) {
         if (!key.getCalculator().supports(node)) {
             return Double.NaN;
         }
@@ -87,7 +88,7 @@ public final class Metrics {
      *
      * @return The value of the metric, or {@code Double.NaN} if the value couln't be computed
      */
-    public static double get(OperationMetricKey key, ASTMethodOrConstructorDeclaration node) {
+    public static double get(MetricKey<OperationMetric> key, ASTMethodOrConstructorDeclaration node) {
         return get(key, node, Version.STANDARD);
     }
 
@@ -101,7 +102,7 @@ public final class Metrics {
      *
      * @return The value of the metric, or {@code Double.NaN} if the value couln't be computed
      */
-    public static double get(OperationMetricKey key, ASTMethodOrConstructorDeclaration node, MetricVersion version) {
+    public static double get(MetricKey<OperationMetric> key, ASTMethodOrConstructorDeclaration node, MetricVersion version) {
         if (!key.getCalculator().supports(node)) {
             return Double.NaN;
         }
@@ -123,7 +124,7 @@ public final class Metrics {
      * @return The value of the metric, or {@code Double.NaN} if the value couln't be computed or {@code option} is
      * {@code null}
      */
-    public static double get(OperationMetricKey key, ASTAnyTypeDeclaration node, ResultOption option) {
+    public static double get(MetricKey<OperationMetric> key, ASTAnyTypeDeclaration node, ResultOption option) {
         return get(key, node, Version.STANDARD, option);
     }
 
@@ -140,7 +141,8 @@ public final class Metrics {
      * @return The value of the metric, or {@code Double.NaN} if the value couln't be computed or {@code option} is
      * {@code null}
      */
-    public static double get(OperationMetricKey key, ASTAnyTypeDeclaration node, MetricVersion version, ResultOption option) {
+    public static double get(MetricKey<OperationMetric> key, ASTAnyTypeDeclaration node, MetricVersion version,
+                             ResultOption option) {
 
         MetricVersion safeVersion = (version == null) ? Version.STANDARD : version;
         return option == null ? Double.NaN