+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/OperationStats.java
@@ -8,9 +8,9 @@ import java.util.HashMap;
 import java.util.Map;
 
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
+import net.sourceforge.pmd.lang.java.oom.api.MetricKey;
 import net.sourceforge.pmd.lang.java.oom.api.MetricVersion;
 import net.sourceforge.pmd.lang.java.oom.api.OperationMetric;
-import net.sourceforge.pmd.lang.java.oom.api.OperationMetricKey;
 
 
 /**
@@ -43,8 +43,8 @@ import net.sourceforge.pmd.lang.java.oom.api.OperationMetricKey;
      *
      * @return The result of the computation, or {@code Double.NaN} if it couldn't be performed
      */
-    /* default */ double compute(OperationMetricKey key, ASTMethodOrConstructorDeclaration node, boolean force,
-                                 MetricVersion version) {
+    /* default */ double compute(MetricKey<OperationMetric> key, ASTMethodOrConstructorDeclaration node,
+                                 boolean force, MetricVersion version) {
 
         ParameterizedMetricKey paramKey = ParameterizedMetricKey.getInstance(key, version);
         Double prev = memo.get(paramKey);