+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/rule/CyclomaticComplexityRule.java
@@ -30,13 +30,13 @@ import net.sourceforge.pmd.lang.rule.properties.IntegerProperty;
  */
 public class CyclomaticComplexityRule extends AbstractJavaMetricsRule {
 
-    public static final IntegerProperty REPORT_LEVEL_DESCRIPTOR = new IntegerProperty(
+    private static final IntegerProperty REPORT_LEVEL_DESCRIPTOR = new IntegerProperty(
         "reportLevel", "Cyclomatic Complexity reporting threshold", 1, 30, 10, 1.0f);
 
-    public static final BooleanProperty REPORT_CLASSES_DESCRIPTOR = new BooleanProperty(
+    private static final BooleanProperty REPORT_CLASSES_DESCRIPTOR = new BooleanProperty(
         "reportClasses", "Add class average violations to the report", true, 2.0f);
 
-    public static final BooleanProperty REPORT_METHODS_DESCRIPTOR = new BooleanProperty(
+    private static final BooleanProperty REPORT_METHODS_DESCRIPTOR = new BooleanProperty(
         "reportMethods", "Add method average violations to the report", true, 3.0f);
 
 
@@ -48,7 +48,7 @@ public class CyclomaticComplexityRule extends AbstractJavaMetricsRule {
         VERSION_MAP.put("ignoreBooleanPaths", CycloMetric.Version.IGNORE_BOOLEAN_PATHS);
     }
 
-    public static final EnumeratedProperty<MetricVersion> CYCLO_VERSION_DESCRIPTOR = new EnumeratedProperty<>(
+    private static final EnumeratedProperty<MetricVersion> CYCLO_VERSION_DESCRIPTOR = new EnumeratedProperty<>(
         "cycloVersion", "Choose a variant of Cyclo or the standard",
         VERSION_MAP, Version.STANDARD, MetricVersion.class, 3.0f);